# public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Queue Management API",
    "description": "API for managing queues and queue entries",
    "version": "1.0.0",
    "contact": {
      "email": "admin@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local API Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/queues/{queue}": {
      "get": {
        "summary": "Get a specific queue",
        "description": "Retrieve details of a specific queue by ID",
        "tags": ["Queues"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "queue",
            "in": "path",
            "required": true,
            "description": "Queue ID",
            "schema": {"type": "integer", "example": 1}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {"type": "integer", "example": 1},
                        "name": {"type": "string", "example": "Customer Service"},
                        "description": {"type": "string", "example": "Main customer service queue"},
                        "status": {"type": "string", "example": "active"},
                        "current_number": {"type": "integer", "example": 5},
                        "created_at": {"type": "string", "format": "date-time"},
                        "updated_at": {"type": "string", "format": "date-time"}
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Queue not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Update a queue",
        "description": "Update an existing queue with new data",
        "tags": ["Queues"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "queue",
            "in": "path",
            "required": true,
            "description": "Queue ID",
            "schema": {"type": "integer", "example": 1}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Updated Customer Service",
                    "description": "Queue name"
                  },
                  "description": {
                    "type": "string",
                    "example": "Updated description",
                    "description": "Queue description"
                  },
                  "status": {
                    "type": "string",
                    "example": "paused",
                    "description": "Queue status"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Queue updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "message": {"type": "string", "example": "Queue updated successfully"},
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {"type": "integer", "example": 1},
                        "name": {"type": "string", "example": "Updated Customer Service"},
                        "description": {"type": "string", "example": "Updated description"},
                        "status": {"type": "string", "example": "paused"},
                        "current_number": {"type": "integer", "example": 5},
                        "created_at": {"type": "string", "format": "date-time"},
                        "updated_at": {"type": "string", "format": "date-time"}
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Queue not found"
          },
          "422": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "summary": "Delete a queue",
        "description": "Delete a queue and all its associated entries",
        "tags": ["Queues"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "queue",
            "in": "path",
            "required": true,
            "description": "Queue ID",
            "schema": {"type": "integer", "example": 1}
          }
        ],
        "responses": {
          "200": {
            "description": "Queue deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "message": {"type": "string", "example": "Queue deleted successfully"}
                  }
                }
              }
            }
          },
          "404": {
            "description": "Queue not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
} 